handle releases better and do build/builddir cleanup (automatic & manual)
we keep releases in the database explicitly.  you have to promote
a build to a release.  releases cannot be deleted.

we automatically removed builds.  only builds that aren't releases
are removed.  the last build for a branch always stays. we only
remove builds when they are not the last 10, or when they are older
than 2 weeks.