improve building under separate user ids
by not requiring a blanket "sudo ..." directive anymore.

now we provide a simple tool "runas", to be run setuid root, that
will execute a command under a different uid.  sudo isn't good
enough, it cannot restrict executing under a range of user id's
(which we need).  also, it is too troublesome to configure.