first bits for privsep code for more securely running code
before this, we always started as an unprivileged process.  then
we would exec a setuid binary to run builds under a unique user id.
we would also need some sudo rules for chowning files.

with this new design, we can start as root. the root process then
forks off an unprivileged child under the ding uid/gid. the two
stay connected through a socket.  the unprivileged ding process
handles all http requests.  when it needs to start a build, it tells
the root process to kick it off.  the root process does so, with
the command running under a unique uid, and sends the file descriptors
for the commands stdout,stderr (and an fd for the status), over the
socket to the unprivileged ding http process.  the http process
then reads the output.

this gets rid of sudo and a setuid program that was written in c.

is a privsep design like this really necessary when using go? it
is supposed to be a "safe" language. the answer is yes. it's a layer
of defense. running all that webserver code as root doesn't feel
right.