improve privsep code.  there is still a bug on macos though.
on macos, if you use the privsep mode, and build under unique uids
that don't have an entry in /etc/passwd, builds will hang.  this
is because the reads on the stdout,stderr from the build command
hang.  linux,openbsd don't show this behaviour.  and switching to
a uid that does have an entry in /etc/passwd also stops showing
this behaviour.  the build command is running, and it writes to its
stdout,stderr: the pipe is full.  the other side is reading.  just
nothing seems to be coming through.

this commit cleans up logging around that part of the code.  and
it switches from os/exec.Command to os.StartProcess.  Command has
requirements about when you can call Wait, which we were probably
violating. StartProcess is easier for us to use.