remove comment about stuff not working
it does work.

the problem lies with xcodebuild that is hanging the "go build"
command. so output was coming in, it's just that the go command was
hanging due to a hanging "xcodebuild -license check"...  and that
only happens when running under a uid without a user.  and seemingly
only on my machine...